requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.39.4
future==0.18.3
h2o==3.40.0.4
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kaggle==1.5.13
kiwisolver==1.4.4
load-dotenv==0.1.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
pywin32==306
pyzmq==25.0.2
requests==2.30.0
six==1.16.0
stack-data==0.6.2
tabulate==0.9.0
text-unidecode==1.3
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
